Move image storage methods to dedicated file
