+++ things/src/main/java/net/bonysoft/doityourselfie/MainActivity.kt
@@ -3,17 +3,10 @@ package net.bonysoft.doityourselfie
 import android.app.Activity
 import android.media.ImageReader
 import android.os.Bundle
-import android.os.Environment
 import android.os.Handler
 import android.os.HandlerThread
 import net.bonysoft.doityourselfie.camera.SelfieCamera
 import net.bonysoft.doityourselfie.camera.dumpFormatInfo
-import timber.log.Timber
-import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import java.text.SimpleDateFormat
-import java.util.Date
 
 class MainActivity : Activity() {
 
@@ -56,34 +49,8 @@ class MainActivity : Activity() {
 
     private fun onPictureTaken(imageBytes: ByteArray?) {
         if (imageBytes != null) {
-            val pictureFile = getOutputMediaFile() ?: return
-            Timber.d("Storing image to ${pictureFile.absolutePath}")
-            try {
-                val fos = FileOutputStream(pictureFile)
-                fos.write(imageBytes)
-                fos.close()
-            } catch (e: IOException) {
-                Timber.e(e, "failed to store to file")
-            }
+            saveImageToFile(imageBytes)
         }
     }
 
-    private fun getOutputMediaFile(): File? {
-        val mediaStorageDir = File(
-            Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES),
-            "MyCameraApp"
-        )
-        if (!mediaStorageDir.exists()) {
-            if (!mediaStorageDir.mkdirs()) {
-                Timber.d("failed to create directory")
-                return null
-            }
-        }
-        // Create a media file name
-        val timeStamp = SimpleDateFormat("yyyyMMdd_HHmmss").format(Date())
-        val mediaFile: File
-        mediaFile = File(mediaStorageDir.path + File.separator + "IMG_" + timeStamp + ".jpg")
-
-        return mediaFile
-    }
 }
